split axis code into separate folders
closes #58
