+++ projects/ngx-d3timeline/src/lib/resources-axis/resources-axis-view-model.ts
