+++ projects/ngx-d3timeline/src/lib/resources-axis/resources-axis.component.ts
