+++ projects/ngx-d3timeline/src/lib/time-axis/time-axis-view-model.ts
