+++ projects/ngx-d3timeline/src/lib/time-axis/time-axis.component.ts
